I moved the REST API generation stuff into it's own module, `autoapi.py`. I am not sure about that name yet but it works for now.
I am also making pushes towards having a general loop for processing REST data,
but also lots of flexibility in how the data is processed. There are two
general methods for input so far: ids or models, and sometimes it's a single
item or a list of items.

The processing will work around the general inputs described above, but it will
also support authentication, customizable status fields and compatibility with
javascript MVC frameworks.