This is the first draft of the new autoapi architecture.
The idea is to have functions that handle the details of operating on a single
structure and then map those across lists of structures.

There are two main components of the autoapi that need to be described:
1. The processing of input
2. The generation of output

Input comes in as either a single id or model dictionary, or a list of ids or
model dictionaries. We validate this input against the model definition and
perform a crud operation on it, like fetching the for an id (GET) or saving a
new copy of the file (PUT).

Output is generated from the results of the crud operation. The crud functions
themselves return lists of two-tuples like this:

    (status description, output data)

This data is then parsed into a configurable representation in HTTP. The
current goal is to be compatible with backbone.js but this effort is still
early.