Added a `render_template` function to the Brubeck class to aid with rendering templates when the `WebMessageHandler` class isn't used.
This function is available only when a `template_loader` function is provided
to a `Brubeck` instance. It is written dynamically and then attached.

It's worth noting this function is unbound, but forms a closure around `self`
and thus has the same access and the same interface as a bound function.

Thanks for bringing this to my attention @kracekumar.