Beginning first wave of refactoring with this commit.
I have renamed `Request` to `Message` and renamed everywhere necessary for consistency.

I also broke out the logic of what a Message is relative to a connection.  They are basically one to one mapped now.  A `Message` doesn't exist yet, but a `WebMessage` does.  Both WSGIMessage and Mongrel2Message derive from here, differing only in the implementation of their parse functions.

This system could be used, in theory, to parse any kind of messages.  In particular, it could easily be an RPC.  Or HTTPRPC.  Or JSONRPC.  It doesn't have to just be web and this commit is pushing us that way.